added health checks to server
