+++ fasjson/web/app.py
@@ -1,16 +1,18 @@
 import os
-import re
 from logging.config import dictConfig
 
-from flask import Flask, jsonify, url_for, current_app
+from flask import Flask
 from werkzeug.routing import BaseConverter
 from werkzeug.exceptions import HTTPException
+from flask_healthz import healthz
 
 from .apis.v1 import blueprint as blueprint_v1
 from .apis.errors import blueprint as blueprint_errors, api as api_errors
 
 from .extensions.flask_gss import FlaskGSSAPI
 from .extensions.flask_ipacfg import IPAConfig
+
+from .base_routes import root
 
 
 class NameConverter(BaseConverter):
@@ -50,6 +52,7 @@ def create_app(config=None):
     # global g variable as described here:
     # https://flask.palletsprojects.com/en/1.1.x/patterns/urlprocessors/#internationalized-blueprint-urls
     app.register_blueprint(blueprint_v1)
+    app.register_blueprint(healthz, url_prefix="/healthz")
 
     # Handler for webserver errors
     app.register_blueprint(blueprint_errors)
@@ -60,24 +63,3 @@ def create_app(config=None):
     app.add_url_rule("/", endpoint="root", view_func=root)
 
     return app
-
-
-def root():
-    blueprints = sorted(
-        [
-            name
-            for name in current_app.blueprints
-            if re.match("^v[0-9]+$", name)
-        ],
-        key=lambda name: int(name[1:]),
-    )
-    apis = [
-        {
-            "version": int(name[1:]),
-            "uri": url_for(f"{name}.root", _external=True),
-            "specs": url_for(f"{name}.specs", _external=True),
-            "docs": url_for(f"{name}.doc", _external=True),
-        }
-        for name in blueprints
-    ]
-    return jsonify({"message": "Welcome to FASJSON", "apis": apis})
