+++ fasjson/web/defaults.cfg
@@ -6,6 +6,10 @@ ERROR_404_HELP = False
 # https://flask-restx.readthedocs.io/en/latest/parsing.html#error-handling
 BUNDLE_ERRORS = True
 
+HEALTHZ = {
+    "live": lambda: None,
+    "ready": "fasjson.web.base_routes.readiness",
+}
 
 # LOGGING = {
 #     "version": 1,
